Move test helper classes to different package
