+++ k9mail-library/src/test/java/com/fsck/k9/mail/helpers/TestMessage.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.mail.transport.mockServer;
+package com.fsck.k9.mail.helpers;
 
 
 import com.fsck.k9.mail.Address;
