+++ k9mail-library/src/test/java/com/fsck/k9/mail/transport/SmtpTransportTest.java
@@ -18,9 +18,9 @@ import com.fsck.k9.mail.internet.MimeMessage;
 import com.fsck.k9.mail.ssl.TrustedSocketFactory;
 import com.fsck.k9.mail.store.StoreConfig;
 import com.fsck.k9.mail.transport.mockServer.MockSmtpServer;
-import com.fsck.k9.mail.transport.mockServer.TestMessage;
+import com.fsck.k9.mail.helpers.TestMessage;
 import com.fsck.k9.mailstore.BinaryMemoryBody;
-import com.fsck.k9.testHelpers.TestTrustedSocketFactory;
+import com.fsck.k9.mail.helpers.TestTrustedSocketFactory;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
